MANPRO-BackendR/python_service/requirements.txt
fastapi
uvicorn
python-multipart
groq
python-dotenv
google-generativeai
